+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/network/node/externalstorage/NetworkNodeExternalStorage.java
@@ -23,6 +23,8 @@ import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
+import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.common.capabilities.CapabilityInject;
 import net.minecraftforge.fluids.FluidStack;
 import net.minecraftforge.fluids.capability.IFluidHandler;
 import net.minecraftforge.items.IItemHandler;
@@ -31,6 +33,11 @@ import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
 
 public class NetworkNodeExternalStorage extends NetworkNode implements IStorageProvider, IGuiStorage, IComparable, IFilterable, IPrioritizable, IType, IAccessType {
+    @CapabilityInject(IDrawerGroup.class)
+    private static final Capability<IDrawerGroup> DRAWER_GROUP_CAPABILITY = null;
+    @CapabilityInject(IDrawer.class)
+    private static final Capability<IDrawer> DRAWER_CAPABILITY = null;
+
     public static final String ID = "external_storage";
 
     private static final String NBT_PRIORITY = "Priority";
@@ -191,20 +198,24 @@ public class NetworkNodeExternalStorage extends NetworkNode implements IStorageP
 
         TileEntity facing = getFacingTile();
 
+        if (facing == null) {
+            return;
+        }
+
         if (type == IType.ITEMS) {
-            if (facing instanceof IDrawerGroup) {
+            if (facing.hasCapability(DRAWER_GROUP_CAPABILITY, getDirection().getOpposite())) {
                 itemStorages.add(new StorageItemDrawerGroup(this, () -> {
                     TileEntity f = getFacingTile();
 
-                    return f instanceof IDrawerGroup ? (IDrawerGroup) f : null;
+                    return (f != null && f.hasCapability(DRAWER_GROUP_CAPABILITY, getDirection().getOpposite())) ? f.getCapability(DRAWER_GROUP_CAPABILITY, getDirection().getOpposite()) : null;
                 }));
-            } else if (facing instanceof IDrawer) {
+            } else if (facing.hasCapability(DRAWER_CAPABILITY, getDirection().getOpposite())) {
                 itemStorages.add(new StorageItemDrawer(this, () -> {
                     TileEntity f = getFacingTile();
 
-                    return f instanceof IDrawer ? (IDrawer) f : null;
+                    return (f != null && f.hasCapability(DRAWER_CAPABILITY, getDirection().getOpposite())) ? f.getCapability(DRAWER_CAPABILITY, getDirection().getOpposite()) : null;
                 }));
-            } else if (facing != null && !(facing.hasCapability(CapabilityNetworkNodeProxy.NETWORK_NODE_PROXY_CAPABILITY, getDirection().getOpposite()) && facing.getCapability(CapabilityNetworkNodeProxy.NETWORK_NODE_PROXY_CAPABILITY, getDirection().getOpposite()).getNode() instanceof IStorageProvider)) {
+            } else if (!(facing.hasCapability(CapabilityNetworkNodeProxy.NETWORK_NODE_PROXY_CAPABILITY, getDirection().getOpposite()) && facing.getCapability(CapabilityNetworkNodeProxy.NETWORK_NODE_PROXY_CAPABILITY, getDirection().getOpposite()).getNode() instanceof IStorageProvider)) {
                 IItemHandler itemHandler = RSUtils.getItemHandler(facing, getDirection().getOpposite());
 
                 if (itemHandler != null) {