+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Refined Storage Changelog
 
 ### 1.5.8
+- Updated Storage Drawers API (raoulvdberge)
 - Fixed bug where disks have to be re-inserted in the Disk Drive in order to work again after rejoining a chunk (raoulvdberge)
 - Autocrafting can now fill water bottles with water from the fluid storage - regular bottles or pattern for regular bottles are required (raoulvdberge)